Distinguish between local and non local errors
The goal of this commit is to introduce Send + Sync
traits to errors wrapped in Error. This conflicts
with local client implementations, that may not
support Send + Sync errors.

To workaround this issue, we introduce a new local
module, in which a local version of Error is put in
place. Local clients will return this error type
instead. Standard and blocking clients will still
return the existing Error type, that will be enriched
with the Send + Sync bounds.